Converted richtext block to a fragment
